outsource clear functions
