+++ Mohr_Circle_legacy/MC_helper_functions.py
@@ -2,6 +2,18 @@
 from math import pi,sqrt,pow,sin,cos,atan 
 
 
+def clear_arrow_source(source_list):
+    empty_dict = dict(xS=[], xE=[], yS=[], yE=[], lW = [])
+    for cds in source_list:
+        ro = len(cds.data["xS"])
+        cds.stream(empty_dict, rollover=-ro)
+
+def clear_rect_source(source_list):
+    empty_dict = dict(x=[], y=[], w=[], h=[], angle=[])
+    for cds in source_list:
+        ro = len(cds.data["x"])
+        cds.stream(empty_dict, rollover=-ro)
+       
 
 
 def calculate_radius_and_center(input_vars):
